[BANK] Fix: 카드 결제시 amount 타입 변경(카드 결제)
Long -> Double

+++ bank/src/main/java/com/goofy/tunabank/v1/dto/card/CardPaymentRequestDto.java
@@ -27,7 +27,7 @@ public class CardPaymentRequestDto {
   private Long merchantId;
 
   @NotNull @Positive
-  private Long paymentBalance;
+  private Double paymentBalance;
 
   @NotEmpty
   private String transactionId;
